==> script2.R <==
################################################################################
# Upload variables
################################################################################

#MERRAclim
bio1_80 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_80s_bio1.tif")
bio1_90 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_90s_bio1.tif")
bio1_00 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_00s_bio1.tif")
bio1 <- mean(bio1_80, bio1_90, bio1_00)
rm(bio1_80, bio1_90, bio1_00)
#
#
bio2_80 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_80s_bio2.tif")
bio2_90 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_90s_bio2.tif")
bio2_00 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_00s_bio2.tif")
bio2 <- mean(bio2_80, bio2_90, bio2_00)
rm(bio2_80, bio2_90, bio2_00)
#
bio3_80 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_80s_bio3.tif")
bio3_90 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_90s_bio3.tif")
bio3_00 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_00s_bio3.tif")
bio3 <- mean(bio3_80, bio3_90, bio3_00)
rm(bio3_80, bio3_90, bio3_00)
#
bio4_80 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_80s_bio4.tif")
bio4_90 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_90s_bio4.tif")
bio4_00 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_00s_bio4.tif")
bio4 <- mean(bio4_80, bio4_90, bio4_00)
rm(bio4_80, bio4_90, bio4_00)
#
bio5_80 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_80s_bio5.tif")
bio5_90 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_90s_bio5.tif")
bio5_00 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_00s_bio5.tif")
bio5 <- mean(bio5_80, bio5_90, bio5_00)
rm(bio5_80, bio5_90, bio5_00)
#
bio6_80 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_80s_bio6.tif")
bio6_90 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_90s_bio6.tif")
bio6_00 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_00s_bio6.tif")
bio6 <- mean(bio6_80, bio6_90, bio6_00)
rm(bio6_80, bio6_90, bio6_00)
#
bio7_80 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_80s_bio7.tif")
bio7_90 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_90s_bio7.tif")
bio7_00 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_00s_bio7.tif")
bio7 <- mean(bio7_80, bio7_90, bio7_00)
rm(bio7_80, bio7_90, bio7_00)
#
bio8_80 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_80s_bio8.tif")
bio8_90 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_90s_bio8.tif")
bio8_00 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_00s_bio8.tif")
bio8 <- mean(bio8_80, bio8_90, bio8_00)
rm(bio8_80, bio8_90, bio8_00)
#
bio9_80 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_80s_bio9.tif")
bio9_90 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_90s_bio9.tif")
bio9_00 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_00s_bio9.tif")
bio9 <- mean(bio9_80, bio9_90, bio9_00)
rm(bio9_80, bio9_90, bio9_00)
#
bio10_80 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_80s_bio10.tif")
bio10_90 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_90s_bio10.tif")
bio10_00 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_00s_bio10.tif")
bio10 <- mean(bio10_80, bio10_90, bio10_00)
rm(bio10_80, bio10_90, bio10_00)
#
bio11_80 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_80s_bio11.tif")
bio11_90 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_90s_bio11.tif")
bio11_00 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_00s_bio11.tif")
bio11 <- mean(bio11_80, bio11_90, bio11_00)
rm(bio11_80, bio11_90, bio11_00)
#
bio12_80 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_80s_bio12.tif")
bio12_90 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_90s_bio12.tif")
bio12_00 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_00s_bio12.tif")
bio12 <- mean(bio12_80, bio12_90, bio12_00)
rm(bio12_80, bio12_90, bio12_00)
#
bio13_80 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_80s_bio13.tif")
bio13_90 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_90s_bio13.tif")
bio13_00 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_00s_bio13.tif")
bio13 <- mean(bio13_80, bio13_90, bio13_00)
rm(bio13_80, bio13_90, bio13_00)
#
bio14_80 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_80s_bio14.tif")
bio14_90 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_90s_bio14.tif")
bio14_00 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_00s_bio14.tif")
bio14 <- mean(bio14_80, bio14_90, bio14_00)
rm(bio14_80, bio14_90, bio14_00)
#
bio15_80 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_80s_bio15.tif")
bio15_90 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_90s_bio15.tif")
bio15_00 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_00s_bio15.tif")
bio15 <- mean(bio15_80, bio15_90, bio15_00)
rm(bio15_80, bio15_90, bio15_00)
#
bio16_80 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_80s_bio16.tif")
bio16_90 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_90s_bio16.tif")
bio16_00 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_00s_bio16.tif")
bio16 <- mean(bio16_80, bio16_90, bio16_00)
rm(bio16_80, bio16_90, bio16_00)
#
bio17_80 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_80s_bio17.tif")
bio17_90 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_90s_bio17.tif")
bio17_00 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_00s_bio17.tif")
bio17 <- mean(bio17_80, bio17_90, bio17_00)
rm(bio17_80, bio17_90, bio17_00)
#
bio18_80 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_80s_bio18.tif")
bio18_90 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_90s_bio18.tif")
bio18_00 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_00s_bio18.tif")
bio18 <- mean(bio18_80, bio18_90, bio18_00)
rm(bio18_80, bio18_90, bio18_00)
#
bio19_80 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_80s_bio19.tif")
bio19_90 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_90s_bio19.tif")
bio19_00 <- rast("D:/Dados climáticos/MERRAclim/10m_mean_00s_bio19.tif")
bio19 <- mean(bio19_80, bio19_90, bio19_00)
rm(bio19_80, bio19_90, bio19_00)
#

#WORLDCLIM
rad1 <- rast("D:/Dados climáticos/WorldClim 2.0/10 minutes/wc2.1_10m_srad/wc2.1_10m_srad_01.tif")
rad2 <- rast("D:/Dados climáticos/WorldClim 2.0/10 minutes/wc2.1_10m_srad/wc2.1_10m_srad_02.tif")
rad3 <- rast("D:/Dados climáticos/WorldClim 2.0/10 minutes/wc2.1_10m_srad/wc2.1_10m_srad_03.tif")
rad4 <- rast("D:/Dados climáticos/WorldClim 2.0/10 minutes/wc2.1_10m_srad/wc2.1_10m_srad_04.tif")
rad5 <- rast("D:/Dados climáticos/WorldClim 2.0/10 minutes/wc2.1_10m_srad/wc2.1_10m_srad_05.tif")
rad6 <- rast("D:/Dados climáticos/WorldClim 2.0/10 minutes/wc2.1_10m_srad/wc2.1_10m_srad_06.tif")
rad7 <- rast("D:/Dados climáticos/WorldClim 2.0/10 minutes/wc2.1_10m_srad/wc2.1_10m_srad_07.tif")
rad8 <- rast("D:/Dados climáticos/WorldClim 2.0/10 minutes/wc2.1_10m_srad/wc2.1_10m_srad_08.tif")
rad9 <- rast("D:/Dados climáticos/WorldClim 2.0/10 minutes/wc2.1_10m_srad/wc2.1_10m_srad_09.tif")
rad10 <- rast("D:/Dados climáticos/WorldClim 2.0/10 minutes/wc2.1_10m_srad/wc2.1_10m_srad_10.tif")
rad11 <- rast("D:/Dados climáticos/WorldClim 2.0/10 minutes/wc2.1_10m_srad/wc2.1_10m_srad_11.tif")
rad12 <- rast("D:/Dados climáticos/WorldClim 2.0/10 minutes/wc2.1_10m_srad/wc2.1_10m_srad_12.tif")

solar_radiation <- mean(rad1, rad2, rad3, rad4, rad5, rad6, rad7, rad8, rad9,
                        rad10, rad11, rad12)
#

rm(rad1, rad2, rad3, rad4, rad5, rad6, rad7, rad8, rad9, rad10, rad11, rad12)

#Human Footprint
h_footprint <- terra::rast("D:/sig/wildareas-v3-2009-human-footprint-geotiff/wildareas-v3-2009-human-footprint.tif")
h_footprint2 <- terra::project(h_footprint, crs(world))

#Create Lists
merraclim_bioclimatic <- list(bio1, bio2, bio3, bio4, bio5, bio6, bio7, bio8, bio9, 
                               bio10, bio11, bio12, bio13, bio14, bio15, bio16, bio17,
                               bio18, bio19)

names(merraclim_bioclimatic) <- c("bio1", "bio2", "bio3", "bio4", "bio5", "bio6", "bio7", 
                                  "bio8", "bio9", "bio10", "bio11", "bio12", "bio13", "bio14", 
                                  "bio15", "bio16", "bio17", "bio18", "bio19")


#EXTRACT
#Extract MerraClim bioclimatic variables
merraclim_bioclimatic_extracted <- data.frame(matrix(nrow = length(final_data_frame_6_SPATIAL)))

for(i in 1:length(merraclim_bioclimatic)){
merraclim_bioclimatic_col <- terra::extract(merraclim_bioclimatic[[i]], final_data_frame_6_SPATIAL)
merraclim_bioclimatic_extracted <- cbind(merraclim_bioclimatic_extracted, merraclim_bioclimatic_col[,2])
}

merraclim_bioclimatic_extracted <- merraclim_bioclimatic_extracted[,-1]
colnames(merraclim_bioclimatic_extracted) <- names(merraclim_bioclimatic) 
#View(merraclim_bioclimatic_extracted)

#Extract solar radiation
solar_radiation_extract <- terra::extract(solar_radiation, final_data_frame_6_SPATIAL)

#Extract human footprint
h_footprint_extract <- terra::extract(h_footprint2, final_data_frame_6_SPATIAL)

final_data_frame_7 <- data.frame(final_data_frame_6_SPATIAL,
           data.frame(terra::crds(final_data_frame_6_SPATIAL)),
           merraclim_bioclimatic_extracted,
           solar_radiation_extract,
           h_footprint_extract)

final_data_frame_7_SPATIAL <- vect(final_data_frame_7, geom=c("x", "y"), 
                                   crs=crs(world), keepgeom=FALSE)
